# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Robot Rangers",
        "description": "Lista de URI's de la API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://robot-rangers-laravel-git-crud-api-robotrangers.vercel.app/"
        }
    ],
    "paths": {
...
        "/rest/pedidos": {
            "get": {
                "tags": [
                    "Prendas"
                ],
                "summary": "Retorna un listado con la informacion de todas las prendas",
                "operationId": "5b29e0b9c2fc7b604daf75a492a63f01",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-05-07 00:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-05-07 00:00:00"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Harden II"
                                                    },
                                                    "marca": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Adidas"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "categoria": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Remeras"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "talle": {
                                                        "type": "string",
                                                        "example": "xl"
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "#FF0000"
                                                    },
                                                    "imagen": {
                                                        "type": "string",
                                                        "example": "https://d2j6dbq0eux0bg.cloudfront.net/images/62219457/3384172981.jpg"
                                                    },
                                                    "precio": {
                                                        "type": "string",
                                                        "example": "9999.99"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Edicion limitada 2023"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Registra la información de un pedido.",
                "operationId": "5cae38cda23f36f8211da02ec7feedba",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "mail_cliente": {
                                                "type": "number"
                                            },
                                            "monto": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "mail_cliente": "clienteiaw2023@gmail.com",
                                    "monto": "9999999999.99"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-05-07 00:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-05-07 00:00:00"
                                        },
                                        "mail_cliente": {
                                            "type": "string",
                                            "example": "clienteiaw2023@gmail.com"
                                        },
                                        "monto": {
                                            "type": "number",
                                            "example": "9999999999.99"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The mail_cliente field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/pedidos/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Retorna la información de un pedido especifico con todas sus prendas",
                "operationId": "e6508c3e918b41350d3abaac0d07fe27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": "1000"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-05-07 20:13:03"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-05-07 20:13:03"
                                                },
                                                "mail_cliente": {
                                                    "type": "string",
                                                    "example": "clienteiaw2023@gmail.com"
                                                },
                                                "monto": {
                                                    "type": "string",
                                                    "example": "35999.99"
                                                },
                                                "prendas": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Harden"
                                                            },
                                                            "cantidad": {
                                                                "type": "number",
                                                                "example": "4"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pedido not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/pedidos/clientes/{mail_cliente}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obtiene una lista con todos los pedidos de un cliente",
                "operationId": "f88b2c61575a4c186cb0622a49b51964",
                "parameters": [
                    {
                        "name": "mail_cliente",
                        "in": "path",
                        "description": "Mail del cliente a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2023-05-07 00:00:00"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2023-05-07 00:00:00"
                                            },
                                            "mail_cliente": {
                                                "type": "string",
                                                "example": "clienteiaw2023@gmail.com"
                                            },
                                            "monto": {
                                                "type": "number",
                                                "example": "9999999999.99"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}